FROMLIST: Decrease cross-threshold roam difficulty
Decrease the roam difficulty when we roam across the short background
scan threshold (i.e. when the current BSS RSSI is below the threshold
and the selected BSS RSSI is above the threshold). This tends to happen
in normal roaming scenarios when we walk from the domain of one AP to
the domain of the other.

BUG=chromium:1014866,chromium:1049782
BUG=b:264617676
TEST=1) network_WiFi_RoamingDiagnostics (WIP test) showed no significant
        decrease in stickiness.
     2) Roaming behavior improved when walking around the building
	(fewer disassocs).

(am from https://patchwork.ozlabs.org/patch/1301986)

Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Signed-off-by: Nicolas Norvez <norvez@chromium.org>
Signed-off-by: Yichen Yu <yichenyu@chromium.org>

TEST=see the following links:
  1. hwsim: https://buganizer.corp.google.com/issues/264617676#comment15
  2. build: https://buganizer.corp.google.com/issues/264617676#comment16
  3. test: https://buganizer.corp.google.com/issues/264617676#comment19

Change-Id: I254f1c0c410ceedbe7fee86ab568c8aad81ce421
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/4210197
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Yichen Yu <yichenyu@google.com>
Auto-Submit: Yichen Yu <yichenyu@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Yichen Yu <yichenyu@google.com>