CHROMIUM: wpa_supplicant: reject auths during explicit roam requests
The roam D-Bus and wpa_cli commands flip the reassociate bit before
calling wpa_supplicant_connect. wpa_supplicant connect eventually aborts
ongoing scans, which causes scan results to be reported. Since the
reassociate bit is set, this will trigger a connection attempt based on
the aborted scan's scan results and cancel the initial connetion
request. This often causes wpa_supplicant to reassociate to the same AP
it is currently associated to.

Add a roam_in_progress flag to indicate that we're currently attempting
to roam via D-Bus so that we don't initate another connection attempt.

BUG=b:179389661
TEST=wifi.RoamDbus with a scan right before roam request

Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Change-Id: I70d358ee1cb9b09c4d91f0eedbb69675a0a65c02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/2683454
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>