FROMLIST: wpa_supplicant: Add support for pregenerated MAC
Add new 'mac_addr' policy (3) with which supplicant expects to also
obtain 'mac_value' with pregenerated value of MAC address to be used for
given SSID.

The main difference between this policy and policy 1 is the ability to
control persistence of the MAC address used.  For example if there is
a requirement to always use the same (but random) MAC address for given
SSID (even if user removes/forgets the network) this could be handled
outside of the wpa_supplicant by using some SSID based hashing scheme to
generate MAC (or by just storing the randomly generated one) and
providing it to wpa_supplicant together with mac_addr=3 policy.

Signed-off-by: Andrzej Ostruszka <amo@semihalf.com>

(am from https://patchwork.ozlabs.org/patch/1553573)

BUG=b:181234131
TEST=Run with updated shill, enable MAC randomization and check if MAC
     is being set to the value obtained from shill.

Change-Id: I9d53dbbdaea22ba62e566bff8b83818980813c27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/3262169
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Joshua Emele <jemele@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>