FROMLIST: Rescan on regulatory domain change
On boot or when resuming after suspending, radios use the lowest common
denominator across all geographies to know which channels/bands to scan.
Because of that, the first scan does not detects APs that are present in
bands that are not allowed everywhere, like the 6GHz band.

With this patch, supplicant triggers a new scan when the regulatory
domain has changed to include channels and bands that were not enabled
previously. It allows the station to connect to WiFi 6E APs faster on
resume in case the only known APs are in the 6GHz band. Without the
patch the station doesn't find any known APs after the first scan, then
it has to wait longer for the next scheduled scan.
If the first scan does contain known APs, the second scan is canceled
either way.

Signed-off-by: Nicolas Norvez <norvez@chromium.org>
(am from https://patchwork.ozlabs.org/patch/1572487/)
(also found at https://marc.info/?i=20211223023358.1645032-1-norvez@chromium.org)

BUG=b:207410721
TEST=cros-workon-${BOARD} start wpa_supplicant-cros
TEST=emerge-${BOARD} wpa_supplicant-cros
TEST=cros deploy ${DUT} wpa_supplicant-cros; reboot ${DUT}
TEST=observe that if the DUT only knows APs in the 6GHz band, the band
reconnects to that AP in ~6-7s after resume (time for 2 scans)
TEST=observe that if the DUT knows AP in the 5GHz band, it reconnects to
that AP after ~3s (one scan) and logs show that the second scan has been
cancelled.

Change-Id: Ieadd967a41668c59fb1021c98087544fb373dd41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/3354086
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>