CHROMIUM: Reject authentication start during BSS TM requests
After receiving a BSS Transition Management request,
wpa_supplicant_connect() will abort ongoing scans, which will cause scan
results to be reported. Since the reassociate bit is set, this will
trigger a connection attempt based on the aborted scan's scan results
and cancel the initial connection request. This often causes
wpa_supplicant to reassociate to the same AP it is currently associated
to instead of the AP it was asked to transition to.

Add a bss_trans_mgmt_in_progress flag to indicate that we're currently
transitioning to a different AP so that we don't initiate another
connection attempt based on the possibly received scan results from a
scan that was in progress at the time the BSS Transition Management
request was received.

This patch is the equivalent of commit
5ac977758d3535e91160ff210f3e8a578031029d for the roaming scenario.

Signed-off-by: Nicolas Norvez <norvez@chromium.org>

BUG=b:212442772
TEST=tast run ${JACUZZI_DUT} wifi.BSSTMRequest

Change-Id: I2c9e5b74528bcc6329fffa34f6a4350da0ef0fd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/3361568
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>