UPSTREAM: Reject authentication start during BSS TM requests
After receiving a BSS Transition Management request,
wpa_supplicant_connect() will abort ongoing scans, which will cause scan
results to be reported. Since the reassociate bit is set, this will
trigger a connection attempt based on the aborted scan's scan results
and cancel the initial connection request. This often causes
wpa_supplicant to reassociate to the same AP it is currently associated
to instead of the AP it was asked to transition to.

Add a bss_trans_mgmt_in_progress flag to indicate that we're currently
transitioning to a different AP so that we don't initiate another
connection attempt based on the possibly received scan results from a
scan that was in progress at the time the BSS Transition Management
request was received.

This is the equivalent of commit 5ac977758d35 ("Reject authentication
start during explicit roam requests") for the roaming scenario.

Signed-off-by: Nicolas Norvez <norvez@chromium.org>
(cherry picked from commit 16b5ea9e9121b6a40e2b4fadbde5894ccbcade09)

BUG=b:212442772
BUG=b:213189089
TEST=See CHROMIUM patch https://crrev.com/c/3361568

Change-Id: I65e745571686aac183f54eb359c31022f023968a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/3383900
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>