FROMLIST: RRM: include passive channels in active beacon report scan
When receiving a beacon report request with the mode set to active,
channels that are marked as NO_IR were not added to the scan request.
However, active mode just mean that active scan is allowed, but not
that it is a must, so these channels should not be omitted.
Include channels that are marked as NO_IR in the scan request even
if the mode is set to active.

Signed-off-by: Avraham Stern <avraham.stern@intel.com>
(am from https://patchwork.ozlabs.org/patch/1600699/)
(also found at https://marc.info/?i=20220303233903.2787228-1-andrei.otcheretianski@intel.com)

BUG=b:222554633
TEST=test_that ${DUT} --args="router_addr=$AP pcap_addr=$PCAP"
suite:wifi_matfunc
TEST=cros_vm --start --board=betty
TEST=cros-workon-betty start wpa_supplicant-cros
TEST=emerge-betty wpa_supplicant-cros
TEST=cros deploy --no-ping localhost:9222 wpa_supplicant-cros
TEST=tast run localhost:9222 wifi.HostapHwsim.full

Signed-off-by: Nicolas Norvez <norvez@chromium.org>
Change-Id: I31e454493ecbbf5333d7e7482fbdf20c79bfd7a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/3511052
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>