UPSTREAM: Fix no_encrypt flag in control port TX for rekeying
The wpa_supplicant check for whether a TK is configured into the driver
was broken during the time this information is needed for rekeying or
reauthenticating with another 4-way handshake. sm->ptk.installed is not
set at the point the EAPOL-Key msg 4/4 is sent and while that means the
initial 4-way handshake needs to prevent encryption, the consecutive
4-way handshake must not be doing that since the old key (TK) is still
in the driver. Fix this so that the EAPOL-Key msg 4/4 during rekeying
does not get transmitted without encryption.

Fixes: a79ed0687197 ("Add no_encrypt flag for control port TX")
Signed-off-by: Jouni Malinen <j@w1.fi>
(cherry picked from commit 98278c0de0cb99ef7725969929fa938ce1c6a154)
Signed-off-by: Kaidong Wang <kaidong@google.com>

BUG=b:256240754
TEST=tast run $DUT wifi.PTK
Tested on asuka and phaser, Message 4 is encrypted and shown as QoS
data in the pcap file.

Change-Id: I2ac65cbd6464cc58531387ef4efae718eb4479b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/4017982
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kaidong Wang <kaidong@google.com>
Tested-by: Kaidong Wang <kaidong@google.com>