FROMLIST: nl80211: Previous MAC isn't a foreign address
When using MAC randomization suppliant can change local MAC address
during roaming scenario:
1. We attach to AP1 (with MAC1).
2. Roaming to AP2 is started:
  a) we send DEAUTH(for AP1, with MAC1)
  b) we change MAC to MAC2 due to randomization
  c) we start authentication for AP2
  d) we get notification about DEAUTH for AP1 (which we ignore)
  e) we complete association with AP2

In point 2d we completely ignore the notification which later causes
problems.

The intended behaviour is as follows: during roaming we generate DEAUTH
(2a) and signal this event right away.  To protect from handling of our
own DEAUTH for the 2nd time supplicant marks 'ignore_next_local_deauth'
variable.  In point 2d we should receive this notification and clear the
flag but this does not happen because MAC1 in the notification is not
the current MAC address (it has been changed in 2b) so this notification
is ignored as a one with a "foreign" address.

So we end up successfully at AP2 but with 'ignore_next_local_deauth'
still set which causes problems.  For example if AP2 shuts down it has
been observed on some drivers that the DEAUTH notification is generated
as a local one and since we have flag to ignore it nothing is reported
over D-Bus.

To address the problem let's store the previously used MAC address and
use it for checking for foreign address (in combination with the current
one).

Signed-off-by: Andrzej Ostruszka <amo@semihalf.com>
(am from https://patchwork.ozlabs.org/patch/1642785/)

BUG=b:230859802
TEST=Run wifi.MARSSIDRoam test on one of boards where it fails with reason
  "context deadline exceeded" (e.g. octopus/blooglet - see stainless)
  and make sure it passes after this fix.
  Also run hwsim tests.

Change-Id: I662b9f8a568d0726a74a0e49c57e2cec9a292b15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/3687749
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>