BACKPORT: Add option to disable SAE key_mgmt without PMF
Add the `sae_check_mfp` global option to limit SAE when PMF will
not be selected for the connection.
With this option SAE is avoided when the hardware is not capable
of PMF due to missing ciphers.
With this option SAE is avoided on capable hardware when the AP
does not enable PMF.

Allows falling back to PSK on drivers with the
WPA_DRIVER_FLAGS_SAE capability but do not support the BIP cipher
necessary for PMF. This enables configurations that can fall back
to WPA-PSK and avoid problems associating with APs configured
with `sae_require_mfp=1`.

Useful when `pmf=1` and `sae_check_mfp=1` are enabled and networks
are configured with ieee80211w=3 (default) and key_mgmt="WPA-PSK SAE".
In this configuration if the device is unable to use PMF due to
lacking BIP group ciphers it will avoid SAE and fallback to
WPA-PSK for that connection.

Signed-off-by: Jeffery Miller <jefferymiller@google.com>
(cherry picked from commit c6f8af507eae658f58de39d15003d3a4f865da1c)

Conflicts:
   wpa_supplicant/wpa_supplicant.c
      Backport conflicts due to additional flags
      WPA_KEY_MGMT_FT_SAE WPA_KEY_MGMT_FT_SAE_EXT_KEY
      that are not in 2.9.

BUG=b:238933377
TEST=connect to wpa2+wpa3 mixed mode AP on the device without BIP
TEST=connect to wpa2+wpa3 mixed mode AP on a device that supports BIP,
 confirm key_mgmt=SAE in `minijail0 -u wpa -g wpa /usr/bin/wpa_cli
 status` output
TEST=tast run -var=wifi.HostapHwsim.runArgs='sae_mixed_mfp
 sae_mixed_check_mfp' betty-dut wifi.HostapHwsim.full

Change-Id: I86655b188645e757794fab9a4feb54c706e931e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/4013703
Commit-Queue: Jeffery Miller <jefferymiller@google.com>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>