CHROMIUM: wpa_supplicant: Apply bias towards 6GHz in roaming
wpa_supplicant_need_to_roam_within_ess applies bias to the minimum
difference in the signal level required to roam. If it is to 2.4GHz,
the bias is 2dBm, if it is to 5GHz, the bias is -2dBm. This bias makes
wpa_supplicant prefer 5GHz over other bands, including 6GHz.

Add a larger bias towards 6GHz, as 6GHz networks usually provide higher
throughput even though the signal is weaker. This CL is marked CHROMIUM
because the current wpa_supplicant branch's roaming algorithm is
different from the upstream, this CL will not be submitted to upstream.

Signed-off-by: Kaidong Wang <kaidong@google.com>

BUG=b:264583988
TEST=emerge-betty wpa_supplicant-cros
TEST=tast -verbose run localhost:9222 wifi.HostapHwsim.validity
TEST=test_that --board=zork host suite:wifi_matfunc

Change-Id: I00143a2db68856401d722b07d6979ab312aba324
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/4146020
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kaidong Wang <kaidong@google.com>
Commit-Queue: Kaidong Wang <kaidong@google.com>