BACKPORT: FROMLIST: Scan 6ghz channels after change to 6ghz-allowed regdom
Drivers will often report regdom changes in the middle of a scan if it
detects during that scan that the regulatory domain has changed. If this
happens and we enter a regdom that supports 6ghz channels when the
previous one didn't (this often happens in 6ghz-capable regdoms for
devices after suspend/resume), immediately trigger a 6ghz-only scan
if we were not able to connect to an AP on a legacy band.

This should significantly improve connection time to 6ghz AP after
regdom has been reset.

Conflicts:
wpa_supplicant/events.c - This involves a partial revert of
crrev/c/3367833 which was never upstreamed

BUG=b:270576915
TEST=manually suspend/resume and observe scan trigger happening
TEST=connect to a 6ghz-only AP and suspend/resume in a loop. Observe
that connection time is reduced from ~15s to <5s. (tested on brya)
TEST=connect to a 5ghz AP and suspend/resume in a loop. Observe that
connection time is not impacted (3s). (tested on brya)
TEST=hwsim tests. All but dbus tests pass (failures are unrelated).
UPSTREAM-TASK=b:287313007

(am from https://patchwork.ozlabs.org/patch/1789843/)

Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Change-Id: I4f644fbb4fc24b1b9f4d6f3dbcfc988cc2caf23f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/4519978
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Kaidong Wang <kaidong@google.com>