FROMLIST: Scan 6ghz channels after change to 6ghz-allowed regdom
Drivers will often report regdom changes in the middle of a scan if it
detects during that scan that the regulatory domain has changed. If this
happens and we enter a regdom that supports 6ghz channels when the
previous one didn't (this often happens in 6ghz-capable regdoms for
devices after suspend/resume), immediately trigger a 6ghz-only scan
if we were not able to connect to an AP on a legacy band.

This should significantly improve connection time to 6ghz AP after
regdom has been reset.

BUG=b:270576915
TEST=no diff with pre-revert tree
UPSTREAM-TASK=b:287313007

(am from https://patchwork.ozlabs.org/patch/1789843/)

Change-Id: I308bdd9292f53d0f75b02c0bce249dafe6635234
Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/4748369
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>