CHROMIUM: apply a bias in roaming/scanning towards 6GHz
Short circuit the roaming and scan result sorting heuristics to apply a
bias towards the 6GHz band. We prefer to be on higher bands due to less
channel utilization and cleaner air. If the SNR is "good enough", stick
to the higher bands.

Note that this is a temporary hack, so we don't plan to send this
upstream. We should eventually come up with a solution for 5GHz as well.

BUG=b:268541533
TEST=manual: sirocco AP configured on all three bands, 6GHz preferred
when device is close, 5GHz preferred when device is far. Walked around
and saw the device roam from 6 -> 5 when AP was far. This differs from
existing behavior in that 6GHz is never preferred, so the device would
rarely ever end up on the 6GHz band.

Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Change-Id: I43de8f125df438073f7606a8b8916d8bc5af7099
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/4617683
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Kaidong Wang <kaidong@google.com>