CHROMIUM: hide roaming and scan sorting logic behind CHROMIUM flag
This includes changes for crrev/c/4617683, crrev/c/4236826,
crrev/c/4210197, and crrev/c/4210195. Note that I'm in the process of
upstreaming parts of these CLs so that we can confine the CHROMIUM block
to just the wpa_supplicant_need_to_roam_within_ess function, but for
now, this should be safe to do to unblock Android from pulling from our
tree.

BUG=b:317253950
TEST=comment out crrev/c/5137788 to make sure that it compiles either
way

Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Change-Id: Ie993c2df2331e2bb50109caf959c7ae303348550
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/5147314
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>