FROMLIST: wpa_supplicant_select_network: check last scan ssids
This recent patch [1] introduced a redundant scan when selecting a
hidden network that was previously scanned and found. This occurs
because the code only checks for the condition
`(wpa_s->no_suitable_network || wpa_s->last_scan_external)`, which
doesn't cover the case where the last scan successfully found the hidden
SSID.

This patch saves the scanned SSIDs from the last scan and updates the
condition to check if the hidden SSID was included.

If the hidden SSID is not found in `last_scan_ssids`, the code checks if
the SSID was found in earlier scan results. If not, it triggers a new
scan. If the SSID is found, a new scan is avoided, resulting in faster
connection times.

[1]
https://w1.fi/cgit/hostap/commit/?id=92374d59d4efea5c8b61ed2ceef141c26bcd7f99

(am from https://patchwork.ozlabs.org/project/hostap/patch/20241105053622.498097-1-arowa@google.com/)

BUG=b:363899693
TEST=Ran hwsim and wifi_matfunc on Rex/Karis/BE200 + NUC-AP

Checked the net.log for wifi.SimpleConnect.hidden5ht20 and there are no
redundant scans.

Ran checkpatch.pl [PASS]

UPSTREAM-TASK=b:372508941

Change-Id: I191713e5eeb88f86c9b838dbbf64168e729b492c
Signed-off-by: Arowa Suliman <arowa@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/5976923
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>