BACKPORT: Check last scan SSIDs before triggering new scan for hidden network
Commit 92374d59d4ef ("Enhance select_network() to trigger new scans in
some cases") introduced a redundant scan when selecting a hidden network
that was previously scanned and found. This occurs because the code only
checks for the condition `(wpa_s->no_suitable_network ||
wpa_s->last_scan_external)`, which doesn't cover the case where the last
scan successfully found the hidden SSID.

Save the scanned SSIDs from the last scan and updates the condition to
check if the hidden SSID was included. If the hidden SSID is not found
in `last_scan_ssids`, the code checks if the SSID was found in earlier
scan results. If not, it triggers a new scan. If the SSID is found, a
new scan is avoided, resulting in faster connection times.

Signed-off-by: Arowa Suliman <arowa@chromium.org>
(cherry picked from commit edd85280ffe43231693ac056e9575aab4a24a2cc)

Conflicts:
   wpa_supplicant/wpa_supplicant_i.h: Removed added local_hw_capab,
   which is already defined under the wpa_supplicant struct.

BUG=372508941
TEST=Ran wifi_matfunc

Change-Id: I46af50ba9f6c61a176c938d37f5ce4bf5f4f8e79
Signed-off-by: Arowa Suliman <arowa@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/6138632
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>