UPSTREAM: Do not remove a currently used BSS entry when removing oldest unknown BSS
wpa_bss_known() might not be sufficient to catch all cases where a BSS
entry is in use. One known example of such a case is OWE transition mode
where the SSID of the transition mode AP is not the same as the one in
the local network profile. Some other cases might exists as well.

If the oldest unknown BSS needs to be removed due to running out of room
in the BSS table and that removed BSS happens to be the currently
associated one, wpa_s->current_bss might become invalid and point to
freed memory. This needs to be avoided to prevent use of freed memory,
so use wpa_bss_in_use() as an extra condition for removing the oldest
unknown BSS.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>

(cherry picked from commit b0f58794420df2fcc195b160e7fd8aa4bf86be2d)
Signed-off-by: David Ruth <druth@chromium.org>

BUG=b:515779476
TEST=Compiles cleanly

Change-Id: I537edc1b7e5003f67adec3e6cd36ac97b215a2b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/8077698
Commit-Queue: David Ruth <druth@chromium.org>
Tested-by: David Ruth <druth@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>